+++ CHANGELOG.rst
@@ -10,6 +10,12 @@ v0.3.0 (24/01/2023)
 
 - Frequency data for ice thickness in data collections (:code:`/dataCollections/{id}/iceThickness`) and autoprocessing programs (:code:`/dataCollections/{id}/iceThickness`)
 
+
+**Changed**
+
+- Moves data collection listing from :code:`/dataCollections` to :code:`/dataGroups/{groupId}/dataCollections`
+- Data collection also displays column with index relative to parent data collection group
+
 ++++++++++
 v0.2.0 (24/01/2023)
 ++++++++++